* Before: Mutant would terminate all worker processes in sequence, while waiting for each pid to terminate separtae. This creates way more context switches than needed and resulted in a serial bottleneck.
* Now mutant signals all worker processes than collects the statuses in a secondary loop.
* Reducing this serial dependency improves mutant exit performance by several seconds on a high core cound machine, and is also measurable on more regular machines.


```
main: time bundle exec mutant run --zombie --since HEAD # 0 mutations selected.

5.69s user 4.35s system 99% cpu 10.061 total

this-branch: bundle exec mutant run --zombie --since HEAD # 0 mutations seleced.

6.23s user 50.38s system 1189% cpu 4.760 total 
```

Note for this benchmarks mutant needs to dynamically self vendor first, `--zombie` is quite expensive and noone but mutant developers should ever use it.